🔀 :: Long type To Integer type

+++ src/main/java/com/msg/after_school/domain/after_school/data/dto/CancleApplyAfterSchoolDto.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor @NoArgsConstructor
 public class CancleApplyAfterSchoolDto {
-    private long afterSchoolId;
+    private Integer afterSchoolId;
 }
