+++ src/main/java/com/msg/after_school/domain/after_school/repository/AfterSchoolRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface AfterSchoolRepository extends JpaRepository<AfterSchool,Long> {
+public interface AfterSchoolRepository extends JpaRepository<AfterSchool, Integer> {
     List<AfterSchool> findAllByIsOpened(Boolean isOpened);
     Optional<AfterSchool> findById(Long Id);
 }
